pyproject.toml
[project]
name = "adfagent"
version = "0.1.0"
description = "ADF Agent - Azure Data Factory 助手"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # CLI and display
    "rich>=14.2.0",
    "prompt-toolkit>=3.0.0",
    # LangChain Agent 框架
    "langchain>=1.0.0",
    "langchain-anthropic>=1.0.0",
    "langgraph>=1.0.0",
    # Configuration
    "python-dotenv>=1.0",
    # Azure SDK
    "azure-identity>=1.15.0",
    "azure-mgmt-datafactory>=4.0.0",
    "azure-mgmt-batch>=17.0.0",
    "azure-keyvault-secrets>=4.7.0",
    "azure-mgmt-resource>=23.0.0",
    "azure-storage-blob>=12.19.0",
    # HTTP requests (for azure_tools)
    "requests>=2.31.0",
]

[project.scripts]
adf_agent = "adf_agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["adf_agent"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
